Fix bug misreporting ERL with external errors in immediate mode
Starting BASIC and hitting Escape would report an "Escape at line 4 billion" as the check of where an error occurred was over strict.
Previously "above PAGE and not in 1st 256 bytes of OUTPUT" now "above &8000", this only lets through a few hundred more bytes anyway.
Change unexpected ERL numbers to be -1 and -2 rather than -1 and 99999.
Tagged as BASIC-1_57-pre4.