Two parsing bug fixes
Lexical.s:
Rework previous fixes (revision 1.5 and revision 1.4) for BPUT#TRACE,<number> confusing the MATCH routine.
While revision 1.5 was better it failed to tokenise IF <expr> GOTO 1234 when the THEN is omitted.
A review of all the keywords shows the only ones that take tokenisable line numbers are TRACE, RESTORE, ON, GOTO, GOSUB and of these only TRACE can be used in a function like manner.
So, only apply the fix from revision 1.5 if it's a TRACE token.
Fix from Martin Avison, fixes ticket #387.
Added a test case in Tests/Match/TokTrace to assist.

Assembler.s:
Only consider SP or LR or PC to be referring to a register name if used in isolation (not as a substring of another word).
While LR and PC aren't particularly common start of words in english, SP is, causing source code with such names to give a "Missing ," error.
We adopt the same logic as for invalid 'R' named variables (eg. r17foo, r9bar) and attempt to find a variable with that name instead.
Fixes ticket #386.

Version 1.59. Tagged as 'BASIC-1_59'