Correction to register vetting
Assembler.s revision 1.18 had a slight flaw in assembling
  LDR Rn,[R1] (or LDMIA Rn,{R1})
would report a bad register because the closing bracket matched the test of being greater than '5' and so skipped the thorough check. So LDR Rn,[R1 ] was OK.
Now, do a more thorough check to determine if the character following could be part of a variable name.

Version 1.59. Not tagged