Assembler encoding of MRS and MSR of banked register
The 'S' of SP was being picked up by CHKPSR as a possible match for SPSR, and setting bit 22 of the output opcode. For banked MRS and MSR this turned
  MRS R9,SP_hyp into MRS R9,SPSR_hyp
  MSR SP_hyp,R1 into MSR SPSR_hyp,R1
rather than change CHKPSR to set bit 22 only when it was sure a PSR was seen, just clear the bit for the banked register case.
Updated TestARMv6v7 to fewer mismatches, and eyeballed the remaining as being presentational only, rather than wrong encoding.
Retagged as BASIC-1_63.