Follow the ARM ARM for VFP opcode VDUP despite a tokens clash
Earlier versions of the VFP assembler extensions cooked up VDPL as a way to dodge the problem of opcode VDUP being tokenised as VDUPRINT when used in its scalar or unconditional form (ie. VDUP.<size> <regs>).
This can be handled with an exception in the tokeniser in a similar manner to MOVEQ and ORR in the conventional ARM instruction set.
VFPLib: change the tables to use VDUP rather than VDPL.
Lexical.s: watch out for specifically VDUP. and tokenise it as VDU followed by a P in memory.
Assembler.s: watch out for token VDU followed by a P and take it to be VDUP.
VFP.s: when matching, allow the token for VDU to expand as though VDU was seen, also, add support for backslash and REM comments (previously only semicolon was allowed).
Tests/VFP/TestVFP: update the test case and 3 other failing tests from earlier fixes.
Tests/Match/TokVDUP: brute force trial of all combinations of VDU and PRINT, to check only VDUP. is caught.

Version 1.68. Tagged as 'BASIC-1_68'