Improve handling of "#0" in VFP/NEON assembler
Where an instruction/encoding only allows a zero value, matching against
the assembly line is performed by looking for an actual "#0" character
sequence. This means that, unlike other locations where literals are
parsed, it's impossible to use a variable, function, or any other kind
of expression to provide the value.

Change the assembler to detect when a "#0" sequence is being looked for,
and use the numeric expression parser instead. This will make it much
easier to use the zero-shift synonym instructions, since it's common for
programs to use expressions to specify/calculate shift amounts (which
may evaluate to zero in some circumstances)

Version 1.80. Tagged as 'BASIC-1_80'